+++ CHANGELOG.md
@@ -20,6 +20,7 @@ Test and beta releases will have experimental functions enabled by default.
 * Migrate on boot completed and display notification migration completed
 * Attempt to fix security exception
 * Moved meta.xml to Meta class
+* Cache permissions
 
 [Open issues](https://github.com/M66B/XPrivacy/issues?state=open)